fix: updates defineOverride
