+++ src/scripts.ts
@@ -1,68 +1,10 @@
 import { sync } from "fast-glob";
-import { compare } from "compare-versions";
-import { resolveResolutions, execPromise, resolveJSON, getRootDeps, updatePackageJSON } from "./utils";
+import { resolveResolutions, execPromise, resolveJSON, updateAppendix, updatePackageJSON } from "./utils";
 import { logger } from "./logger";
 import { IS_DEBUGGING } from "./constants";
-import {
-  Appendix,
-  Options,
-  OverridesType,
-  UpdateAppendixOptions,
-} from "./interfaces";
+import { Appendix, Options, OverridesType } from "./interfaces";
 
 const log = logger({ file: "scripts.ts", isLogging: IS_DEBUGGING });
-
-export async function updateAppendix({
-  debug = false,
-  dependencies,
-  resolutions,
-  name,
-  version,
-  appendix = {},
-  exec = execPromise,
-}: UpdateAppendixOptions): Promise<Appendix> {
-  const dependencyList = Object.keys(dependencies);
-  const resolutionsList = Object.keys(resolutions);
-  try {
-    const resolutionRootDeps = await getRootDeps({ resolutions: resolutionsList, debug, exec });
-    const updatedAppendix = resolutionsList.reduce(
-      (acc: Appendix, resolution: string): Appendix => {
-        if (dependencyList.includes(resolution)) {
-          const hasResolutionOverride = compare(
-            resolutions[resolution],
-            dependencies[resolution],
-            ">"
-          );
-          if (hasResolutionOverride) {
-            const key = `${resolution}@${resolutions[resolution]}`;
-            const { rootDeps = [] } = resolutionRootDeps.find((dep) => dep.resolution === resolution) || {};
-            return {
-              ...appendix,
-              ...acc,
-              [key]: {
-                dependents: {
-                  ...appendix?.[key]?.dependents,
-                  ...acc?.[key]?.dependents,
-                  [name]: version,
-                },
-                ...(rootDeps.length ? { rootDeps } : {}),
-              },
-            };
-          }
-        }
-        return acc || {};
-      },
-      {}
-    );
-    log.debug("updateAppendix:fn:", {
-      updatedAppendix,
-    });
-    return updatedAppendix;
-  } catch (err) {
-    log.error('updatedAppendix:fn', { error: err });
-    return appendix;
-  }
-}
 
 export async function update(options: Options): Promise<Appendix | void> {
   const {
@@ -87,9 +29,11 @@ export async function update(options: Options): Promise<Appendix | void> {
     const { dependencies = {}, name, version } = currentPackageJSON;
     const dependenciesList = Object.keys(dependencies);
     if (!dependenciesList.length) return acc;
+    // TODO this is currently only mapping to the resolution object
     const hasOverriddenDependencies = dependenciesList.some((dependencyItem) =>
       resolutionsList.includes(dependencyItem)
     );
+    console.log({ hasOverriddenDependencies });
     if (!hasOverriddenDependencies) return acc;
     const appendixItem = await updateAppendix({
       debug,
