+++ src/utils/getRootDeps.ts
@@ -1,0 +1,33 @@
+
+import { execPromise } from "./index";
+import { GetRootDeps, RootDepItem } from "../interfaces";
+
+import { logger } from "../logger";
+import { IS_DEBUGGING } from "../constants";
+
+const log = logger({ file: "utils.ts", isLogging: IS_DEBUGGING });
+
+export async function getRootDeps({ resolutions, exec = execPromise }: GetRootDeps): Promise<Array<RootDepItem>> {
+  const rootDepsList = Promise.all(
+    resolutions.map(async (resolution: string): Promise<RootDepItem> => {
+      try {
+        const runner = 'npm'
+        const cmd = ['ls', resolution, '--json']
+        const { dependencies } = await exec(runner, cmd);
+        const rootDeps = Object.keys(dependencies).map((dependency) => `${dependency}@${dependencies[dependency].version}`);
+        log.debug(`getRootDeps: ${resolution} has direct dependendents: ${rootDeps.join(", ")}`);
+        return {
+          resolution,
+          rootDeps
+        };
+      } catch (err) {
+        log.error('getRootDeps:', { error: err });
+        return {
+          resolution,
+          rootDeps: []
+        };
+      }
+    })
+  );
+  return rootDepsList;
+}
